+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.12] - 2019-01-08
+### Added
+- Added solarium/solarium dependency
+- Added solr configuration and index functionality
+- Added `solr_port` & `solr_collection` configuration
+- Added `cms:solr:index` command that indexes pages
+- Added `solr_serializer` option to `page_templates` configuration
+- Added serializer interface `RevisionTen\CMS\Interface\SolrSerializerInterface`
+- Added IndexService and SearchService
+- Added example solr collection `conf` folder (solr 6 compatible) in Resources/solr_conf_example
+- Added fulltext search to controller element
+
 ## [1.4.11] - 2019-01-08
 ### Added
 - Added `multiple` and `expanded` option to DoctrineType